+++ src/components/v2/common/ReactSelect.utils.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { ReactComponent as ArrowDown } from '../assets/icons/ic-chevron-down.svg'
 import { components } from 'react-select'
 import Tippy from '@tippyjs/react'
-import { noop, stopPropagation } from '@devtron-labs/devtron-fe-common-lib'
+import { multiSelectStyles, noop, stopPropagation } from '@devtron-labs/devtron-fe-common-lib'
 import { Environment } from '../../cdPipeline/cdPipeline.types'
 import { CLUSTER_TERMINAL_MESSAGING } from '../../ClusterNodes/constants'
 
@@ -287,4 +287,30 @@ export const groupHeaderStyle = {
         backgroundColor: 'var(--N100)',
         marginBottom: 0,
     }),
+}
+
+export const groupStyle = () => {
+  return {
+      ...multiSelectStyles,
+      menu: (base) => ({ ...base, zIndex: 9999, textAlign: 'left' }),
+      control: (base) => ({ ...base, border: '1px solid #d6dbdf', width: '450px' }),
+      group: (base) => ({
+          ...base,
+          paddingTop: 0,
+          paddingBottom: 0,
+      }),
+      groupHeading: (base) => ({
+          ...base,
+          fontWeight: 600,
+          fontSize: '12px',
+          height: '28px',
+          color: 'var(--N900)',
+          backgroundColor: 'var(--N100)',
+          marginBottom: 0,
+      }),
+      indicatorsContainer: (provided, state) => ({
+          ...provided,
+      }),
+      option: getCustomOptionSelectionStyle(),
+  }
 }
