+++ src/css/base.scss
@@ -2036,6 +2036,10 @@ button.anchor {
     max-width: 1000px;
 }
 
+.dc__mxw-180 {
+    max-width: 180px;
+}
+
 .dc__mxw-155 {
     max-width: 155px;
 }
